Separate helpers from variant interfaces
