+++ src/environment/Command.h
@@ -66,25 +66,4 @@ namespace ymwm::environment::commands {
                                MoveToNextGroup,
                                MoveToPrevGroup>;
 
-  template <std::size_t Index =
-                std::variant_size_v<environment::commands::Command> - 1ul>
-  static inline std::optional<environment::commands::Command>
-  try_find_command(std::string_view command_type) noexcept {
-    if constexpr (0ul <= Index) {
-      using CommandType =
-          std::variant_alternative_t<Index, environment::commands::Command>;
-      if (command_type == CommandType::type) {
-        return CommandType{};
-      }
-
-      if constexpr (0ul == Index) {
-        return std::nullopt;
-      } else {
-        return try_find_command<Index - 1ul>(command_type);
-      }
-
-    } else {
-      return std::nullopt;
-    }
-  }
 } // namespace ymwm::environment::commands
